requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile dashboard/pyproject.toml -o requirements.txt
altair==5.4.1
    # via
    #   dashboard (dashboard/pyproject.toml)
    #   streamlit
atpublic==5.0
    # via ibis-framework
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
beautifulsoup4==4.12.3
    # via nbconvert
bleach==6.1.0
    # via nbconvert
blinker==1.8.2
    # via streamlit
cachetools==5.5.0
    # via
    #   google-auth
    #   streamlit
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via streamlit
defusedxml==0.7.1
    # via nbconvert
duckdb==1.1.1
    # via dashboard (dashboard/pyproject.toml)
fastjsonschema==2.20.0
    # via nbformat
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
google-api-core==2.21.0
    # via google-api-python-client
google-api-python-client==2.149.0
    # via dashboard (dashboard/pyproject.toml)
google-auth==2.35.0
    # via
    #   dashboard (dashboard/pyproject.toml)
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.2.0
    # via
    #   dashboard (dashboard/pyproject.toml)
    #   google-api-python-client
google-auth-oauthlib==1.2.1
    # via dashboard (dashboard/pyproject.toml)
googleapis-common-protos==1.65.0
    # via google-api-core
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
ibis-framework==9.5.0
    # via dashboard (dashboard/pyproject.toml)
idna==3.10
    # via requests
jinja2==3.1.4
    # via
    #   altair
    #   nbconvert
    #   pydeck
jsonschema==4.23.0
    # via
    #   altair
    #   nbformat
jsonschema-specifications==2024.10.1
    # via jsonschema
jupyter-client==8.6.3
    # via nbclient
jupyter-core==5.7.2
    # via
    #   jupyter-client
    #   nbclient
    #   nbconvert
    #   nbformat
jupyterlab-pygments==0.3.0
    # via nbconvert
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   nbconvert
mdurl==0.1.2
    # via markdown-it-py
mistune==3.0.2
    # via nbconvert
narwhals==1.10.0
    # via altair
nbclient==0.10.0
    # via nbconvert
nbconvert==7.16.4
    # via dashboard (dashboard/pyproject.toml)
nbformat==5.10.4
    # via
    #   dashboard (dashboard/pyproject.toml)
    #   nbclient
    #   nbconvert
numpy==2.1.2
    # via
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
oauthlib==3.2.2
    # via requests-oauthlib
packaging==24.1
    # via
    #   altair
    #   nbconvert
    #   streamlit
pandas==2.2.3
    # via streamlit
pandocfilters==1.5.1
    # via nbconvert
parsy==2.1
    # via ibis-framework
pathlib==1.0.1
    # via dashboard (dashboard/pyproject.toml)
pillow==10.4.0
    # via streamlit
platformdirs==4.3.6
    # via jupyter-core
polars==1.10.0
    # via dashboard (dashboard/pyproject.toml)
proto-plus==1.24.0
    # via google-api-core
protobuf==5.28.2
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
    #   streamlit
pyarrow==17.0.0
    # via streamlit
pyarrow-hotfix==0.6
    # via dashboard (dashboard/pyproject.toml)
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydeck==0.9.1
    # via streamlit
pygments==2.18.0
    # via
    #   nbconvert
    #   rich
pyparsing==3.2.0
    # via httplib2
python-dateutil==2.9.0.post0
    # via
    #   ibis-framework
    #   jupyter-client
    #   pandas
python-dotenv==1.0.1
    # via dashboard (dashboard/pyproject.toml)
pytz==2024.2
    # via
    #   ibis-framework
    #   pandas
pyzmq==26.2.0
    # via jupyter-client
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   google-api-core
    #   requests-oauthlib
    #   streamlit
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rich==13.9.3
    # via streamlit
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   bleach
    #   python-dateutil
smmap==5.0.1
    # via gitdb
soupsieve==2.6
    # via beautifulsoup4
sqlglot==25.20.2
    # via ibis-framework
streamlit==1.39.0
    # via dashboard (dashboard/pyproject.toml)
tenacity==9.0.0
    # via streamlit
tinycss2==1.3.0
    # via nbconvert
toml==0.10.2
    # via streamlit
toolz==0.12.1
    # via ibis-framework
tornado==6.4.1
    # via
    #   jupyter-client
    #   streamlit
traitlets==5.14.3
    # via
    #   jupyter-client
    #   jupyter-core
    #   nbclient
    #   nbconvert
    #   nbformat
typing-extensions==4.12.2
    # via
    #   altair
    #   ibis-framework
    #   streamlit
tzdata==2024.2
    # via pandas
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.2.3
    # via requests
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
